requirements.txt
beautifulsoup4==4.9.3
Django==2.2.12
django-braces==1.14.0
django-crispy-forms==1.9.1
importlib-metadata==2.0.0
mysql-connector-python==8.0.21
mysqlclient==2.0.1
Pillow==8.0.1
protobuf==3.13.0
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
zipp==3.3.0
